requirements.txt
Django==2.1.3
psycopg2==2.7.6
gunicorn==19.9.0
whitenoise==4.1
python-decouple==3.1
dj-database-url==0.5.0
djangorestframework==3.9.0
pytest==3.10.0
pytest-django==3.4.3
model_mommy==1.6.0
django-filter==2.0.0
parameterized==0.6.1
pytest-cov==2.6.0
sentry_sdk==0.5.4
drf-yasg==1.11.0
djangorestframework-jwt==1.11.0
requests==2.20.1
requests-cache==0.4.13